+++ Kernel/Net/UDPSocket.h
@@ -42,7 +42,7 @@ private:
     static Lockable<HashMap<u16, UDPSocket*>>& sockets_by_port();
 
     virtual int protocol_receive(const KBuffer&, void* buffer, size_t buffer_size, int flags) override;
-    virtual int protocol_send(const void*, int) override;
+    virtual int protocol_send(const void*, size_t) override;
     virtual KResult protocol_connect(FileDescription&, ShouldBlock) override;
     virtual int protocol_allocate_local_port() override;
     virtual KResult protocol_bind() override;
